python-312-perf-profiler/requirements.txt
Pillow==9.5.0
pyinstrument==4.5.0
